対象日記情報取得関数を共通化対応
